// src/types/representative/personal-data/representativeUpdateFinancial.ts
export type RepresentativeUpdateFinancial = {
  pixPrimaryValue: string
  pixPrimaryType: string
  pixSecondaryValue: string
  pixSecondaryType: string
  bankPrimaryCode: string
  bankPrimaryAgency: string
  bankPrimaryAccount: string
  bankSecondaryAgency: string
  bankSecondaryAccount: string
}